Move the make code from psc into the purescript library
